+++ apps/questions/views.py
@@ -7,6 +7,13 @@ from .models import Question
 from .serializers import QuestionSerializer
 
 
+def question_post(request, post_id):
+    question_object = Question.objects.get(id=post_id)
+    question_object.views = question_object.views + 1
+    question_object.save()
+    return question_object
+
+
 @api_view(["POST"])
 def post_question(request):
     serializer = QuestionSerializer(
@@ -41,5 +48,12 @@ def question_view(request, id):
             return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     elif request.method == "DELETE":
-        question.delete()
-        return Response({"message": "The question was deleted"})
+        if question.user == request.user:
+            question.delete()
+            return Response(
+                {"message": "The question was deleted"},
+                status=status.HTTP_204_NO_CONTENT,
+            )
+        return Response(
+            {"message": "Unauthorized"}, status=status.HTTP_401_UNAUTHORIZED
+        )
